% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_codons_to_aa.R
\name{dna_codons_to_aa}
\alias{dna_codons_to_aa}
\title{Translates 3-character codon to 1-character amino acid}
\usage{
dna_codons_to_aa(codons)
}
\arguments{
\item{codons}{input DNA codon}
}
\description{
Translates 3-character codon to 1-character amino acid
}
