% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_codons.R
\name{mk_codons}
\alias{mk_codons}
\title{Pulls out substrings from position s}
\usage{
mk_codons(dna, s = 1)
}
\arguments{
\item{dna}{input dna string}

\item{s}{start postition}
}
\description{
Pulls out substrings from position s
}
