% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Returns random DNA on l nucleotides}
\usage{
random_dna(l)
}
\arguments{
\item{l}{number of wanted nucleotides in the DNA}
}
\description{
Returns random DNA on l nucleotides
}
