% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rna.R
\name{random_rna}
\alias{random_rna}
\title{Creates l random nucleotides from rna}
\usage{
random_rna(l)
}
\arguments{
\item{l}{number of chosen nucleotides}
}
\description{
Creates l random nucleotides from rna
}
